backend/pyproject.toml
[tool.poetry]
name = "pycon-quiz"
version = "0.1.0"
description = ""
authors = ["Marco Acierno <marcoaciernoemail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.4"
django-environ = "^0.4.5"
django-model-utils = "^4.0.0"
django-fsm = "^2.7.0"
django-choices = "^1.7.1"
django-object-actions = "^2.0.0"
django-nested-admin = "^3.2.4"
django-admin-interface = "^0.12.1"
uvicorn = "0.11.5"
django-cors-headers = "^3.2.1"
aioredis = "^1.3.1"
strawberry-graphql = "^0.25.1"
psycopg2 = "^2.8.5"
django-seal = "^1.2.3"
isort = "^4.3.21"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pdbpp = "^0.10.2"
django-debug-toolbar = "^2.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
